+++ CHANGELOG.md
@@ -3,7 +3,9 @@ CHANGELOG for Sulu
 
 * dev-master
     * ENHANCEMENT #3698 [ContentBundle]           SEO description length changed from 155 to 320
-    * HOTFIX      #3695 [MediaBundle]             Keep doctrine 2.5
+    * HOTFIX      #3695 [RouteBundle]             Added check for empty request format
+    * HOTFIX      #3695 [WebsiteBundle]           Added check for empty request format
+    * HOTFIX      #3696 [MediaBundle]             Keep doctrine 2.5
 
 * 1.6.11 (2017-12-13)
     * HOTFIX      #3690 [ContentBundle]           Fix saving of not yet started text editor